feature: added yaml loaders as plugin instead of default adapter
